move record-related models into models/record.py
